Weekly Challenge for week 6
Fixed it up nice and good. Aligned the difficulty buttons to be centred and changed their colour to match the difficulty they're going to be. The food correctly destroys on mouse click. The score displays correctly, adding up with each bit of food clicked. The restart button shows when the game over happens. The difficulty buttons alter the spawnrate of the food so that easy is easier and hard is harder.